feat: write tests for writing out IPv4 and IPv6 packets

+++ examples/write_ipv4_udp.rs
@@ -33,7 +33,7 @@ fn main() {
         //time to live
         20,
         //contained protocol is udp
-        IpNumber::Udp,
+        IpNumber::Udp as u8,
         //source ip address
         [192,168,1,42],
         //destination ip address
